+++ Formula/netductor.rb
@@ -1,28 +1,28 @@
 class Netductor < Formula
   desc "Netductor control plane CLI / TUI"
   homepage "https://github.com/PavelNeyman/netductor"
-  version "0.8.75"
+  version "0.8.76"
   license "MIT"
 
   on_macos do
     on_arm do
-      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.75/netductor-darwin-arm64"
-      sha256 "bbb5c958c07c0927aaaf0fc3a3808bb5fe0420cb760c5cdf7e50df38f6fac7be"
+      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.76/netductor-darwin-arm64"
+      sha256 "5f96139d07f651bed28b4ef5896fc79810e6e0e98942a572e68af6e25da0a078"
     end
     on_intel do
-      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.75/netductor-darwin-amd64"
-      sha256 "9451bf54a66524cc4937bb639b5363eb5cc7bd3d4864778227b7ac0b24bb6e0b"
+      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.76/netductor-darwin-amd64"
+      sha256 "8eaaad47ae2469a580b9e9cd7bc82f6663af57ada39e9cf8985fd9fb82f46702"
     end
   end
 
   on_linux do
     on_intel do
-      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.75/netductor-linux-amd64"
-      sha256 "715fce796ce5cf6b6f7b0a77b71d5d8af790fc9f68b0e75d848b712b138368d1"
+      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.76/netductor-linux-amd64"
+      sha256 "933c228252132e657905f535c2c6ffbb10b3208f5c984efa57d7ec72adedc330"
     end
     on_arm do
-      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.75/netductor-linux-arm64"
-      sha256 "11a8d49248b0bd01fa224b8a4b3415c3c5eeb3a24dcd782563e2073bf1e23aea"
+      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.76/netductor-linux-arm64"
+      sha256 "3e9357ac010aba68496a804624a7d1f480c7aa610f1195dc6a41e86ee869f163"
     end
   end